Remove Vundle stuff and update default colorscheme
The vundle stuff is too hard to use in offline cases which happens for
me.  Probably better just to store the stuff I want in /etc/vim
directory so I don't have to rely on an internet connection.

Similarly, set the default colorscheme to a built-in one so that
I don't have to worry about it not being there.  If I get solarized
added to /etc/vim I may change the default back.